requirements.txt
fastapi
uvicorn
langchain
langgraph
sqlalchemy
python-docx
pillow
requests
python-dotenv
